Extract types to their own files
