+++ src/Features/Core/Portable/IncrementalCaches/SymbolTreeInfoIncrementalAnalyzer.cs
@@ -1,0 +1,197 @@
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+using System;
+using System.Collections.Concurrent;
+using System.Collections.Generic;
+using System.Diagnostics;
+using System.Linq;
+using System.Threading;
+using System.Threading.Tasks;
+using Microsoft.CodeAnalysis.FindSymbols;
+using Microsoft.CodeAnalysis.Remote;
+using Microsoft.CodeAnalysis.SolutionCrawler;
+using Roslyn.Utilities;
+
+namespace Microsoft.CodeAnalysis.IncrementalCaches
+{
+    internal partial class SymbolTreeInfoIncrementalAnalyzerProvider
+    {
+        private class SymbolTreeInfoIncrementalAnalyzer : IncrementalAnalyzerBase
+        {
+            private readonly ConcurrentDictionary<ProjectId, SymbolTreeInfo> _projectToInfo;
+            private readonly ConcurrentDictionary<string, MetadataInfo> _metadataPathToInfo;
+
+            public SymbolTreeInfoIncrementalAnalyzer(
+                ConcurrentDictionary<ProjectId, SymbolTreeInfo> projectToInfo,
+                ConcurrentDictionary<string, MetadataInfo> metadataPathToInfo)
+            {
+                _projectToInfo = projectToInfo;
+                _metadataPathToInfo = metadataPathToInfo;
+            }
+
+            public override async Task AnalyzeDocumentAsync(Document document, SyntaxNode bodyOpt, InvocationReasons reasons, CancellationToken cancellationToken)
+            {
+                if (!SupportAnalysis(document.Project))
+                {
+                    return;
+                }
+
+                if (bodyOpt != null)
+                {
+                    // This was a method body edit.  We can reuse the existing SymbolTreeInfo if
+                    // we have one.  We can't just bail out here as the change in the document means
+                    // we'll have a new checksum.  We need to get that new checksum so that our
+                    // cached information is valid.
+                    if (_projectToInfo.TryGetValue(document.Project.Id, out var cachedInfo))
+                    {
+                        var checksum = await SymbolTreeInfo.GetSourceSymbolsChecksumAsync(
+                            document.Project, cancellationToken).ConfigureAwait(false);
+
+                        var newInfo = cachedInfo.WithChecksum(checksum);
+                        _projectToInfo[document.Project.Id] = newInfo;
+                        return;
+                    }
+                }
+
+                await UpdateSymbolTreeInfoAsync(document.Project, cancellationToken).ConfigureAwait(false);
+            }
+
+            public override Task AnalyzeProjectAsync(Project project, bool semanticsChanged, InvocationReasons reasons, CancellationToken cancellationToken)
+            {
+                if (!SupportAnalysis(project))
+                {
+                    return Task.CompletedTask;
+                }
+
+                return UpdateSymbolTreeInfoAsync(project, cancellationToken);
+            }
+
+            private async Task UpdateSymbolTreeInfoAsync(Project project, CancellationToken cancellationToken)
+            {
+                Debug.Assert(SupportAnalysis(project));
+
+                // Produce the indices for the source and metadata symbols in parallel.
+                var tasks = new List<Task>
+                {
+                    InvokeAsync(project, (self, project, _, cancellationToken) => self.UpdateSourceSymbolTreeInfoAsync(project, cancellationToken), null, cancellationToken),
+                    InvokeAsync(project, (self, project, _, cancellationToken) => self.UpdateReferencesAsync(project, cancellationToken), null, cancellationToken)
+                };
+
+                await Task.WhenAll(tasks).ConfigureAwait(false);
+            }
+
+            private async Task UpdateSourceSymbolTreeInfoAsync(Project project, CancellationToken cancellationToken)
+            {
+                var checksum = await SymbolTreeInfo.GetSourceSymbolsChecksumAsync(project, cancellationToken).ConfigureAwait(false);
+                if (!_projectToInfo.TryGetValue(project.Id, out var projectInfo) ||
+                    projectInfo.Checksum != checksum)
+                {
+                    projectInfo = await SymbolTreeInfo.GetInfoForSourceAssemblyAsync(
+                        project, checksum, cancellationToken).ConfigureAwait(false);
+
+                    Contract.ThrowIfNull(projectInfo);
+                    Contract.ThrowIfTrue(projectInfo.Checksum != checksum, "If we computed a SymbolTreeInfo, then its checksum much match our checksum.");
+
+                    // Mark that we're up to date with this project.  Future calls with the same 
+                    // semantic version can bail out immediately.
+                    _projectToInfo[project.Id] = projectInfo;
+                }
+            }
+
+            [PerformanceSensitive("https://github.com/dotnet/roslyn/issues/36158", AllowCaptures = false, Constraint = "Avoid captures to reduce GC pressure when running in the host workspace.")]
+            private Task InvokeAsync(Project project, Func<SymbolTreeInfoIncrementalAnalyzer, Project, PortableExecutableReference, CancellationToken, Task> func, PortableExecutableReference reference, CancellationToken cancellationToken)
+            {
+                var isRemoteWorkspace = project.Solution.Workspace.Kind == WorkspaceKind.RemoteWorkspace;
+                return isRemoteWorkspace
+                    ? GetNewTask(this, func, project, reference, cancellationToken)
+                    : func(this, project, reference, cancellationToken);
+
+                static Task GetNewTask(SymbolTreeInfoIncrementalAnalyzer self, Func<SymbolTreeInfoIncrementalAnalyzer, Project, PortableExecutableReference, CancellationToken, Task> func, Project project, PortableExecutableReference reference, CancellationToken cancellationToken)
+                {
+                    return Task.Run(() => func(self, project, reference, cancellationToken), cancellationToken);
+                }
+            }
+
+            [PerformanceSensitive("https://github.com/dotnet/roslyn/issues/36158", AllowCaptures = false)]
+            private Task UpdateReferencesAsync(Project project, CancellationToken cancellationToken)
+            {
+                // Process all metadata references. If it remote workspace, do this in parallel.
+                var tasks = new List<Task>();
+
+                foreach (var reference in project.MetadataReferences.OfType<PortableExecutableReference>())
+                {
+                    tasks.Add(
+                        InvokeAsync(project, (self, project, reference, cancellationToken) => self.UpdateReferenceAsync(project, reference, cancellationToken), reference, cancellationToken));
+                }
+
+                return Task.WhenAll(tasks);
+            }
+
+            private async Task UpdateReferenceAsync(
+                Project project, PortableExecutableReference reference, CancellationToken cancellationToken)
+            {
+                var key = GetReferenceKey(reference);
+                if (key == null)
+                {
+                    return;
+                }
+
+                var checksum = SymbolTreeInfo.GetMetadataChecksum(project.Solution, reference, cancellationToken);
+                if (!_metadataPathToInfo.TryGetValue(key, out var metadataInfo) ||
+                    metadataInfo.SymbolTreeInfo.Checksum != checksum)
+                {
+                    var info = await SymbolTreeInfo.GetInfoForMetadataReferenceAsync(
+                        project.Solution, reference, checksum, loadOnly: false, cancellationToken: cancellationToken).ConfigureAwait(false);
+
+                    Contract.ThrowIfNull(info);
+                    Contract.ThrowIfTrue(info.Checksum != checksum, "If we computed a SymbolTreeInfo, then its checksum much match our checksum.");
+
+                    // Note, getting the info may fail (for example, bogus metadata).  That's ok.  
+                    // We still want to cache that result so that don't try to continuously produce
+                    // this info over and over again.
+                    metadataInfo = new MetadataInfo(info, metadataInfo.ReferencingProjects ?? new HashSet<ProjectId>());
+                    _metadataPathToInfo[key] = metadataInfo;
+                }
+
+                // Keep track that this dll is referenced by this project.
+                metadataInfo.ReferencingProjects.Add(project.Id);
+            }
+
+            public override Task RemoveProjectAsync(ProjectId projectId, CancellationToken cancellationToken)
+            {
+                _projectToInfo.TryRemove(projectId, out _);
+                RemoveMetadataReferences(projectId);
+
+                return Task.CompletedTask;
+            }
+
+            private void RemoveMetadataReferences(ProjectId projectId)
+            {
+                foreach (var kvp in _metadataPathToInfo.ToArray())
+                {
+                    if (kvp.Value.ReferencingProjects.Remove(projectId))
+                    {
+                        if (kvp.Value.ReferencingProjects.Count == 0)
+                        {
+                            // This metadata dll isn't referenced by any project.  We can just dump it.
+                            _metadataPathToInfo.TryRemove(kvp.Key, out _);
+                        }
+                    }
+                }
+            }
+
+            private bool SupportAnalysis(Project project)
+            {
+                if (!project.SupportsCompilation)
+                {
+                    // Not a language we can produce indices for (i.e. TypeScript).  Bail immediately.
+                    return false;
+                }
+
+                return RemoteFeatureOptions.ShouldComputeIndex(project.Solution.Workspace);
+            }
+        }
+    }
+}
